Refactor AppBar into its own component
