+++ src/components/ApplicationBar.jsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+import AppBar from 'material-ui/AppBar';
+import IconButton from 'material-ui/IconButton';
+
+import LINKS from '../constants/links';
+
+export default class ApplicationBar extends React.Component {
+  render() {
+    return (
+      <AppBar
+        title="WebRTC communication portal"
+        showMenuIconButton={false}
+        iconElementRight={
+          <IconButton iconClassName="material-icons" href={LINKS.GITHUB_LINK} >
+            code
+          </IconButton >
+        }
+      />
+    );
+  }
+}
