+++ src/constants/links.js
@@ -1,3 +1,5 @@
-const GIHTUB_LINK = 'https://github.com/VaclavStebra/webrtc-communication-portal';
+const GITHUB_LINK = 'https://github.com/VaclavStebra/webrtc-communication-portal';
 
-export default GIHTUB_LINK;
+export default {
+  GITHUB_LINK
+};
